+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 ## 0.0.17
 
 - Added RoleBasedGuard
-- Added RoleBasedGuard to /settings/student-type/add
+- Added RoleBasedGuard to add student type page
+- Move user auth verify to AuthBased Guard
+- Added toastify for error alert on AuthGuard and RoleBaseGuard
+- Fix unneccessary padding bottom on navbar menu icon
+- Added breadcrumb component for all the pages
+- Fix routes paths for module, settings, student type and student type add
 
 ## 0.0.16